moved TicketOrder.vue from components to desktop folder
